# docs/swagger.yaml
basePath: /api
definitions:
...
  models.AuditLog:
    properties:
      aksi:
        type: string
      detail:
        type: string
      id:
        type: integer
      timestamp:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.LostDocument:
    properties:
      created_at:
        type: string
      id:
        type: integer
      last_updated_by:
        $ref: '#/definitions/models.User'
      last_updated_by_id:
        description: LastUpdatedByID adalah pengguna yang login dan melakukan aksi Update
          terakhir
        type: integer
      lokasi_hilang:
        type: string
      lost_items:
        items:
          $ref: '#/definitions/models.LostItem'
        type: array
      nomor_surat:
        type: string
      operator:
        $ref: '#/definitions/models.User'
      operator_id:
        description: |-
          --- FIELD BARU UNTUK AUDIT ---
          Operator adalah pengguna yang login dan melakukan aksi Create
        type: integer
      pejabat_persetuju:
        $ref: '#/definitions/models.User'
      pejabat_persetuju_id:
        type: integer
      petugas_pelapor:
        $ref: '#/definitions/models.User'
      petugas_pelapor_id:
        description: Petugas yang namanya akan tercetak di surat
        type: integer
      resident:
        $ref: '#/definitions/models.Resident'
      resident_id:
        type: integer
      status:
        type: string
      tanggal_laporan:
        type: string
      tanggal_persetujuan:
        type: string
      updated_at:
        type: string
    type: object
  models.LostItem:
    properties:
      deskripsi:
        type: string
      id:
        type: integer
      lost_document_id:
        type: integer
      nama_barang:
        type: string
    type: object
  models.Resident:
    properties:
      agama:
        type: string
      alamat:
        type: string
      created_at:
        type: string
      id:
        type: integer
      jenis_kelamin:
        type: string
      nama_lengkap:
        type: string
      nik:
        type: string
      pekerjaan:
        type: string
      tanggal_lahir:
        type: string
      tempat_lahir:
        type: string
      updated_at:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      jabatan:
        description: KANIT SPKT, ANGGOTA JAGA REGU
        type: string
      nama_lengkap:
        type: string
      nrp:
        type: string
      pangkat:
        type: string
      peran:
        description: SUPER_ADMIN, OPERATOR
        type: string
      regu:
        description: I, II, III
        type: string
      updated_at:
        type: string
    type: object
  services.ChartDataDTO:
    properties:
      data:
        items:
          type: integer
        type: array
      labels:
        items:
          type: string
        type: array
    type: object
  services.DashboardStatsDTO:
    properties:
      active_users:
        type: integer
      docs_monthly:
        type: integer
      docs_today:
        type: integer
      docs_yearly:
        type: integer
    type: object
  services.PieChartDataDTO:
    properties:
      background_colors:
        items:
          type: string
        type: array
      data:
        items:
          type: integer
        type: array
      labels:
        items:
          type: string
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: yusuf.aar@gmail.com
    name: MUHAMMAD YUSUF ABDURROHMAN
    url: https://github.com/dope-s/simdokpol-go
  description: Ini adalah dokumentasi API untuk aplikasi Sistem Informasi Manajemen
    Dokumen Kepolisian.
  license:
    name: MIT License
    url: https://github.com/dope-s/simdokpol-go/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: SIMDOKPOL API
  version: "1.0"
paths:
  /audit-logs:
    get:
      description: Mengambil seluruh riwayat aktivitas yang tercatat di sistem. Hanya
        bisa diakses oleh Super Admin.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AuditLog'
            type: array
        "500":
          description: 'Error: Gagal mengambil data log audit'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mendapatkan Semua Log Audit
      tags:
      - Audit Log
  /audit-logs/export:
    get:
      description: Mengunduh seluruh riwayat log audit sebagai file .xlsx. Hanya bisa
        diakses oleh Super Admin.
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: File Excel (.xlsx)
          schema:
            type: file
        "500":
          description: 'Error: Gagal membuat file ekspor'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Ekspor Log Audit ke Excel
      tags:
      - Audit Log
...
  /notifications/expiring-documents:
    get:
      description: Mengambil daftar dokumen yang akan segera masuk masa arsip untuk
        pengguna yang sedang login.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LostDocument'
            type: array
      security:
      - BearerAuth: []
      summary: Mendapatkan Notifikasi Dokumen Kedaluwarsa
      tags:
      - Dashboard & Stats
...
  /stats:
    get:
      description: Mengambil data statistik utama untuk kartu-kartu di dasbor (laporan
        hari ini, bulan ini, tahun ini, dan total pengguna).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.DashboardStatsDTO'
        "500":
          description: 'Error: Gagal mengambil data statistik'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mendapatkan Statistik Utama Dasbor
      tags:
      - Dashboard & Stats
  /stats/item-composition:
    get:
      description: Mengambil data label dan jumlah untuk grafik pai komposisi barang
        hilang di dasbor.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.PieChartDataDTO'
        "500":
          description: 'Error: Gagal mengambil data komposisi barang'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mendapatkan Data Komposisi Barang
      tags:
      - Dashboard & Stats
  /stats/monthly-issuance:
    get:
      description: Mengambil data label dan jumlah penerbitan surat per bulan untuk
        grafik area di dasbor.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.ChartDataDTO'
        "500":
          description: 'Error: Gagal mengambil data grafik'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mendapatkan Data Grafik Bulanan
      tags:
      - Dashboard & Stats
...
securityDefinitions:
  BearerAuth:
    description: Masukkan token JWT Anda dengan format 'Bearer {token}'.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"